Json handling

+++ GwMaster/GwMasterCtrl_v5.0.X/src/lib/jsmn.h
@@ -40,12 +40,12 @@ enum jsmnerr {
  */
 typedef struct {
 	jsmntype_t type;
-	int start;
-	int end;
-	int size;
-   int index;
+	short start;
+	short end;
+	short size;
+   short index;
 #ifdef JSMN_PARENT_LINKS
-	int parent;
+	short parent;
 #endif
 } jsmntok_t;
